Allow code to keep multiple caches, split out some logic to other files
