+++ src/config.py
@@ -1,12 +1,11 @@
-import sys
+import logging
+import random
 
-expname = 'exp'
-if len(sys.argv) > 3:
-    expname = sys.argv[3]
+seed = 45
+random.seed(seed)
 
-thresh = 75
-if len(sys.argv) > 4:
-    thresh = int(sys.argv[4])
+expname = 'data'
+thresh = 35
 
 addition = '-' + expname + '-' + str(thresh)
 outputDir = 'output/'  # All files in this folder will be deleted
@@ -14,3 +13,14 @@ outputDirRaw = outputDir + 'raw/'
 outputDirDist = outputDir + 'dist/'
 outputDirStats = 'stats/'
 outputDirFigs = outputDir + 'figs' + addition
+
+sequentialDistanceCacheName = f'data/sequentialDistance-{thresh}-{seed}.pkl'
+statisticalDistanceCacheName = f'data/statisticalDistance-{thresh}-{seed}.pkl'
+propertiesCacheName = f'data/normalizedProperties-{thresh}-{seed}.pkl'
+pklCache = f'data/meta-{thresh}-{seed}.pkl'
+mappingCache = f'data/mapping-{thresh}-{seed}.pkl'
+totalLabelsCache = f'data/totalLabels-{thresh}-{seed}.pkl'
+
+logLevel = logging.INFO
+generateTSNEGraphs = False
+generateAllGraphs = True
